fix(nvim,json): move json config to separate module
It used to be in web, it's cleaner to have it by itself. I also thought
about getting it to work with local schema files, but never got that
far. I need to figure out how I can load a schema for a file without
having it match on the filename.