refactor: attempt at refactoring the ansible code to be cleaner
I started with just updating the Windows WSL config. It currently
doesn't work because I have changed the git names to be nested.

Getting nested names to work wasn't very nice... I wish there was a
better way of doing it. However, I think the result will be cleaner in
the end.

I will continue this refactoring on another system.